+++ R/build-manual.R
@@ -10,16 +10,22 @@ build_manual <- function(pkg = ".", path = NULL) {
   pkg <- as.package(pkg)
   path <- path %||% path_dir(pkg$path)
   name <- paste0(pkg$package, "_", pkg$version, ".pdf", collapse = " ")
-  tryCatch(msg <- callr::rcmd("Rd2pdf", cmdargs = c(
-    "--force",
-    paste0("--output=", path, "/", name),
-    pkg$path
-  ), fail_on_status = TRUE, spinner = FALSE),
-  error = function(e) {
-    cat(e$stdout)
-    cli::cli_abort(c("x" = no_wrap(e$stderr),
-                     "!" = "Failed to build manual"))
-  })
+  tryCatch(
+    msg <- callr::rcmd(
+      "Rd2pdf",
+      cmdargs = c(
+        "--force",
+        paste0("--output=", path, "/", name),
+        pkg$path
+      ),
+      fail_on_status = TRUE,
+      spinner = FALSE
+    ),
+    error = function(e) {
+      cat(e$stdout)
+      cli::cli_abort(c("x" = no_wrap(e$stderr), "!" = "Failed to build manual"))
+    }
+  )
 
   cat(msg$stdout)
   invisible(msg)